Removed encoding.go
