+++ log.go
@@ -2,12 +2,15 @@ package raft
 
 import (
 	"bufio"
+	"encoding/binary"
 	"io"
 	"io/fs"
 	"os"
 	"path/filepath"
 
 	"github.com/jmsadair/raft/internal/errors"
+	pb "github.com/jmsadair/raft/internal/protobuf"
+	"google.golang.org/protobuf/proto"
 )
 
 // Log represents the internal component of Raft that is responsible
@@ -103,6 +106,59 @@ func NewLogEntry(index uint64, term uint64, data []byte, entryType LogEntryType)
 // Two log entries are considered conflicting if they have the same index but different terms.
 func (e *LogEntry) IsConflict(other *LogEntry) bool {
 	return e.Index == other.Index && e.Term != other.Term
+}
+
+func encodeLogEntry(w io.Writer, entry *LogEntry) error {
+	pbEntry := &pb.LogEntry{
+		Index:     entry.Index,
+		Term:      entry.Term,
+		Data:      entry.Data,
+		Offset:    entry.Offset,
+		EntryType: pb.LogEntry_LogEntryType(entry.EntryType),
+	}
+
+	buf, err := proto.Marshal(pbEntry)
+	if err != nil {
+		return err
+	}
+
+	size := int32(len(buf))
+	if err := binary.Write(w, binary.BigEndian, size); err != nil {
+		return err
+	}
+
+	if _, err := w.Write(buf); err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func decodeLogEntry(r io.Reader) (LogEntry, error) {
+	var size int32
+	if err := binary.Read(r, binary.BigEndian, &size); err != nil {
+		return LogEntry{}, err
+	}
+
+	buf := make([]byte, size)
+	if _, err := io.ReadFull(r, buf); err != nil {
+		return LogEntry{}, err
+	}
+
+	pbEntry := &pb.LogEntry{}
+	if err := proto.Unmarshal(buf, pbEntry); err != nil {
+		return LogEntry{}, err
+	}
+
+	entry := LogEntry{
+		Index:     pbEntry.GetIndex(),
+		Term:      pbEntry.GetTerm(),
+		Data:      pbEntry.GetData(),
+		Offset:    pbEntry.GetOffset(),
+		EntryType: LogEntryType(pbEntry.EntryType),
+	}
+
+	return entry, nil
 }
 
 // persistentLog implements the Log interface. Not concurrent safe.
